+++ hutool-core/src/main/java/cn/hutool/core/collection/CollUtil.java
@@ -47,10 +47,13 @@ import cn.hutool.core.util.StrUtil;
 import cn.hutool.core.util.TypeUtil;
 
 /**
- * 集合相关工具类，包括数组
+ * 集合相关工具类<p>
+ * 此工具方法针对{@link Collection}及其实现类封装的工具。<p>
+ * 由于{@link Collection} 实现了{@link Iterable}接口，因此部分工具此类不提供，而是在{@link IterUtil} 中提供
  * 
  * @author xiaoleilu
  * @since 3.1.1
+ * @see IterUtil
  */
 public class CollUtil {